# swagger.yaml
openapi: 3.0.3
info:
  title: Notes API - OpenAPI 3.0
  version: 1.0.1
  description: |-
    This is a sample services based on the OpenAPI 3.0 specification. You can make short notes(404 symbols)
tags:
  - name: user
    description: Operations about user
  - name: notes
    description: Operations about note
  - name: tag
    description: Operations about tag
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: New user was successfully created!
        '400':
          description: The username is used by other user
        '404':
          description: The email is used by other user
  
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
          example: theUser
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied

  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      description: ''
      operationId: logoutUser
      parameters: []
      responses:
        default:
          description: successful operation
  /user/{user_id}:
    get:
      tags:
        - user
      summary: Get user by user id
      description: ''
      operationId: getUserByid
      parameters:
        - name: user_id
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
          example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Invalid user_id supplied
        '404':
          description: User not found
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          description: user that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: successful operation
        '400':
          description: User with such username already exists.
        '404':
          description: User with such email already exists.
      security:
        - note_auth:
          - admin
          - user
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User was deleted.
        '400':
          description: Invalid user_id supplied
        '404':
          description: An user with provided ID was not found.
      security:
        - note_auth:
          - admin
          - user

  /notes:
    post:
      tags:
        - notes
      summary: Create note
      description: This can only be done by the logged in user.
      operationId: createNote
      requestBody:
        description: Created note object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesCreate'
          application/xml:
            schema:
              $ref: '#/components/schemas/NotesCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NotesCreate'
      responses:
        '200':
          description: Note was added successfully
        '400':
          description: This title has already exist.
        '404':
          description: Can't use the same tag again
      security:
        - note_auth:
          - admin
          - user
  /notes/{note_id}:
    get:
      tags:
        - notes
      summary: Get note by id
      description: ''
      operationId: getNoteById
      parameters:
        - name: note_id
          in: path
          description: 'The name that needs to be fetched. Use 1 for testing. '
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteInfo'
            application/xml:
              schema:
                $ref: '#/components/schemas/NoteInfo'
        '404':
          description: An Note with provided ID was not found.
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - notes
      summary: Update note
      description: This can only be done by the logged in user.
      operationId: updateNotes
      parameters:
        - name: note_id
          in: path
          description: id that need to be deleted
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent note
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
          application/xml:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NotesUpdate'
      responses:
        '200':
          description: Note was updated successfully
        '404':
          description: An Note with provided ID was not found.
        '400':
          description: Note with such title already exists.
      security:
        - note_auth:
          - admin
          - user

    delete:
      tags:
        - notes
      summary: Delete note
      description: This can only be done by the logged in user.
      operationId: deleteNotes
      parameters:
        - name: note_id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Note was deleted.
        '404':
          description: Note not found
      security:
        - note_auth:
          - admin
          - user

  /tag:
    post:
      tags:
        - tag
      summary: Create tag
      description: This can only be done by the logged in user.
      operationId: createTag
      requestBody:
        description: Created tag object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: New tag was successfully created!
        '400':
          description: Tag with such name already exists
        '500':
          description: Invalid data
      security:
        - note_auth:
          - admin
          - user
  /tag/{id}:
    get:
      tags:
        - tag
      summary: Get tag by id
      description: ''
      operationId: getTagById
      parameters:
        - name: id
          in: path
          description: 'The name that needs to be fetched. Use 1 for testing. '
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid id supplied
        '404':
          description: Note not found
      security:
        - note_auth:
          - admin
          - user
    put:
      tags:
        - tag
      summary: Update tag
      description: This can only be done by the admin
      operationId: updateTag
      parameters:
        - name: id
          in: path
          description: id that need to be deleted
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent note
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: The tag was successfully updated
        '404':
          description: An tag with provided ID was not found.
        '400':
          description: Tag with such number already exists
          
      security:
        - note_auth:
          - admin
    delete:
      tags:
        - tag
      summary: Delete tag
      description: This can only be done by the admin
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tag was deleted.
        '404':
          description: An tag with provided ID was not found.
      security:
        - note_auth:
          - admin

  /notes/allowed:
    post:
      tags:
        - notes
      summary: allow a editing
      description: This can only be done by the logged in user.
      operationId: allowEditing
      requestBody:
        description: Created note object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/allowedNote'
          application/xml:
            schema:
              $ref: '#/components/schemas/allowedNote'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/allowedNote'
      responses:
        '200':
          description: New editor was added.
        '400':
          description: An Note with provided ID was not found.
        '404':
          description: An User with provided ID was not found.
      security:
        - note_auth:
          - admin
          - user

components:
  schemas:
    UserInfo:
      type: object
      required:
      - email
      properties:
        firstName:
          type: string
          example: Ivan
        lastName:
          type: string
          example: James
        username:
          type: string
          example: theUser
        email:
          type: string
          example: ivan@email.com
        numOfNotes:
          type: integer
          format: int64
          example: 9
        dateOfCreating:
          type: string
          format: date
          example: '2022-10-02'
        numOfEditingNotes:
          type: integer
          format: int64
          example: 9
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NoteInfo'
    UserCreate:
      required:
      - email
      - password
      - username
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
          example: theUser
        email:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
          example: theUser
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
      xml:
        name: user
    NotesCreate:
      type: object
      required:
        - id
        - ownerId
        - isPublic
      properties:
        ownerId:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: 'My title'
        isPublic:
          type: boolean
          example: true
        text:
          type: string
          example: 'Today i was in Lviv. It was great!'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        
      xml:
        name: notes
    NoteInfo:
      type: object
      required:
        - id
        - ownerId
        - isPublic
      properties:
        ownerId:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          example: 'My title'
        isPublic:
          type: boolean
          example: true
        text:
          type: string
          example: 'Today i was in Lviv. It was great!'
        dateOfEditing:
          type: string
          format: date
          example: '2022-10-02'  
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        editors:
          type: array
          items:
              properties:
                username:
                  type: string
                  example: theUser
                email:
                  type: string
                  example: John@let.com
            
        
      xml:
        name: notes
    NotesUpdate:
      type: object
      properties:
        title:
          type: string
          example: 'My title'
        isPublic:
          type: boolean
          example: true
        text:
          type: string
          example: 'Today i was in Lviv. It was great!'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
      xml:
        name: notesUp
    Tag:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: '#lviv'
      xml:
        name: tag
    allowedNote:
      type: object
      properties:
        note_id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 2
      xml:
        name: allowedNotes
  requestBodies:
    UserArray:
      description: List of user object
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/UserCreate'
  securitySchemes:
    note_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: url
          scopes:
            admin : can do all actions
            user : can do only allowed action
    api_key:
      type: apiKey
      name: api_key
      in: header